Added Arduino Data Transfer Code
The Arduino code forwards 12 byte data packets from the Serial port (at 115200 baud) to the SNES when the SNES requests the data using the latch pin. The last bit of the 2nd byte in the  transfered data packet is 1 if new data was read from PC before the SNES requested new data , 0 otherwise (indicates new data).